core/db_lib/requirements.txt
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
faiss-cpu>=1.7.4
psutil>=5.9.0
